Move Hologram component to complex scene example
