+++ hotham/src/components/mod.rs
@@ -5,7 +5,6 @@ pub mod global_transform;
 pub mod grabbable;
 pub mod hand;
 pub mod hmd;
-pub mod hologram;
 pub mod info;
 pub mod joint;
 pub mod local_transform;
@@ -27,7 +26,6 @@ pub use global_transform::GlobalTransform;
 pub use grabbable::Grabbable;
 pub use hand::Hand;
 pub use hmd::HMD;
-pub use hologram::Hologram;
 pub use info::Info;
 pub use joint::Joint;
 pub use local_transform::LocalTransform;
